fix compilation error on older gcc

+++ gemrb/core/Scriptable/Actor.cpp
@@ -3903,7 +3903,7 @@ void Actor::dump(StringBuffer& buffer) const
 	buffer.appendFormatted("Area:       %.8s   ", Area );
 	buffer.appendFormatted("Dialog:     %.8s\n", Dialog );
 	buffer.appendFormatted("Global ID:  %d   PartySlot: %d\n", GetGlobalID(), InParty);
-	buffer.appendFormatted("Script name:%.32s    Current action: %d    Total: %ld\n", scriptName, CurrentAction ? CurrentAction->actionID : -1, actionQueue.size());
+	buffer.appendFormatted("Script name:%.32s    Current action: %d    Total: %ld\n", scriptName, CurrentAction ? CurrentAction->actionID : -1, (long) actionQueue.size());
 	buffer.appendFormatted("Int. Flags: 0x%x ", InternalFlags);
 	buffer.appendFormatted("TalkCount:  %d   ", TalkCount );
 	buffer.appendFormatted("Allegiance: %d   current allegiance:%d\n", BaseStats[IE_EA], Modified[IE_EA] );
